// generated/src/app/api-models/claim-api-header-profsvc-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimApiHeaderProfsvcW {
  claimIdNo: string;
  priorAuthNumber: string;
  secondaryAuth: string;
  subscriberId: string;
  personNumber: string;
  patDateOfBirth: string;
  patSex: string;
  patLastName: string;
  patFirstName: string;
  patMi: string;
  socialSecNo: string;
  providerNumber: string;
  vendorId: string;
  referProvIdNo: string;
  diagnosisCode1: string;
  diagnosisCode2: string;
  diagnosisCode3: string;
  diagnosisCode4: string;
  totalClaimCharge: string;
  userDefined1: string;
  userDefined2: string;
  otherInsuranceInd: string;
  assignOfBenefits: string;
  emplRelatedInd: string;
  accidentInd: string;
  accidentSymptomDate: string;
  patControlNo: string;
  provAssignInd: string;
  facilityLabIdNo: string;
  sameSimilarSymptomInd: string;
  sameSimilarSymptomDate: string;
  admissionDate1: string;
  dischargeDate1: string;
  batchNumber: string;
  status: string;
  authWaiveMethod: string;
  providerPanel: string;
  providerIpa: string;
  authClass: string;
  refProvPcp: string;
  unableToWorkBeginDate: string;
  unableToWorkThruDate: string;
  reservedLocalUseHdr: string;
  outsideLabInd: string;
  outsideLabCharges: string;
  medicaidResubmitCode: string;
  originalReferenceNumber: string;
  patientPaidAmount: string;
  facilityNumber: string;
  labNumber: string;
  autoAccidentInd: string;
  autoAccidentState: string;
  releaseMedRecsInd: string;
  onsetTime: string;
  ocCarrierCode: string;
  cobAllwdAmtRule: string;
  cobAllwdAmtRsn: string;
  paySubscriber: string;
  diamondId: string;
  providerPostalCode: string;
  explanationAttached: string;
}